Move the encoding into a different file.
